# app/openapi.yaml
openapi: 3.1.0
info:
  title: Domeo No-Code Calculators — Doors API
  version: "1.0.0"
  description: |
    OpenAPI для пилота **Doors** (Next 14). Включает актуализированный контракт **медиа-аплоада**:
    `POST /api/admin/media/upload` — multipart (`model` + `file[]`), сохранение как `public/assets/doors/encodeURIComponent(model).ext`.
servers:
  - url: http://localhost:3000
    description: Dev (Next 14)
  - url: https://example.dev.domeo.app
    description: Remote Dev (CI remote-smoke)
...
security:
  - bearerAuth: []
paths:
...
  /api/catalog/doors/options:
    get:
      tags: [Catalog]
      operationId: getDoorsOptions
      summary: Получить опции каталога (dependsOn-цепочка)
      parameters:
        - in: query
          name: style
          schema: { type: string, nullable: true }
        - in: query
          name: model
          schema: { type: string, nullable: true }
        - in: query
          name: finish
          schema: { type: string, nullable: true }
        - in: query
          name: color
          schema: { type: string, nullable: true }
        - in: query
          name: type
          schema: { type: string, nullable: true }
        - in: query
          name: width
          schema: { type: string, nullable: true }
        - in: query
          name: height
          schema: { type: string, nullable: true }
      responses:
        "200":
          description: Опции с учетом выбранных фильтров
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoorsOptionsResponse"
  /api/price/doors:
    post:
      tags: [Pricing]
      operationId: priceDoors
      summary: Расчёт цены (auto-pricing)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceRequest"
      responses:
        "200":
          description: Цена рассчитана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceResponse"
        "400":
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/cart/export/doors/factory:
    post:
      tags: [Cart/Export]
      operationId: exportDoorsFactory
      summary: Экспорт заказа на фабрику (stub v1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        "200":
          description: Успешно (stub v1)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponseFactory"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          nullable: true
      required: [error]
...
    DoorsOptionsResponse:
      type: object
      properties:
        style:
          type: array
          items: { type: string }
        model:
          type: array
          items: { type: string }
        finish:
          type: array
          items: { type: string }
        color:
          type: array
          items: { type: string }
        type:
          type: array
          items: { type: string }
        width:
          type: array
          items: { type: string }
        height:
          type: array
          items: { type: string }
        kits:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
              price_rrc: { type: number }
        handles:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
              price_rrc:
                type: number
                nullable: true
              price_opt:
                type: number
                nullable: true
              price_group_multiplier:
                type: number
                nullable: true
    PriceRequest:
      type: object
      description: Параметры конфигуратора для расчёта
      properties:
        model: { type: string }
        finish: { type: string }
        color: { type: string }
        type: { type: string }
        width: { type: string }
        height: { type: string }
        hardware_kit_id:
          type: string
          nullable: true
        handle_id:
          type: string
          nullable: true
        qty:
          type: integer
          minimum: 1
          default: 1
      required: [model, finish, color, type, width, height]
    PriceResponse:
      type: object
      properties:
        ok: { type: boolean }
        base_price: { type: number }
        handle_price:
          description: Цена ручки (по правилу ручки из спецификаций)
          type: number
          nullable: true
        total_price: { type: number }
        currency: { type: string, default: RUB }
      required: [ok, base_price, total_price]
    ExportRequest:
      type: object
      description: |
        **stub v1** — допускает произвольные поля. Будет ужесточён в v1.1 (ExportDoorsRequest).
      additionalProperties: true
    ExportResponseBase:
      type: object
      properties:
        ok:
          type: boolean
          const: true
        type:
          type: string
          enum: [kp, invoice, factory]
        received:
          type: object
          additionalProperties: true
      required: [ok, type, received]
...
    ExportResponseFactory:
      allOf:
        - $ref: "#/components/schemas/ExportResponseBase"
        - type: object
          properties:
            type:
              type: string
              const: factory
...